[x86_64] Tons of changes of long and ulong for their safe types uint32_t, int32_t, etc. Added -Wall and -O0 flags when compiling in DEBUG mode

+++ tuxcap/demo3/GameApp.cpp
@@ -262,17 +262,17 @@ void GameApp::LoadingThreadCompleted()
 	mAlteredImg = (DDImage*) CopyImage(mTurbotImg);
 
 	// 2. Now we need to get the pixel data. The pixel data is stored as
-	// an unsigned long array, where each entry represents the RGBA value.
+	// an uint32_t array, where each entry represents the RGBA value.
 	// The data is actually stored in ARGB format, where alpha is
 	// the leftmost byte and blue is the rightmost byte.
-	unsigned long* bits = ((DDImage*)mAlteredImg)->GetBits();
+	uint32_t* bits = ((DDImage*)mAlteredImg)->GetBits();
 
 	// 3. Now we will loop over each pixel in the image. The size of the bits array
 	// is simply the width times the height.
 	for (int i = 0; i < mAlteredImg->GetWidth() * mAlteredImg->GetHeight(); i++)
 	{
 		// 4. Get the ARGB color value for this pixel
-		unsigned long c = bits[i];
+		uint32_t c = bits[i];
 
 		// 5. To illustrate the ARGB storage format, we will assign each
 		// component to a variable, although we're actually only going to care
@@ -287,7 +287,7 @@ void GameApp::LoadingThreadCompleted()
 		// Let's alter these to produce a grayscale image using one of many
 		// conversion methods. This method uses 30% of the red value,
 		// 59% of the green value, and 11% of the blue value:
-		unsigned long gray = (unsigned long) ((float)red * 0.30f + (float)green * 0.59f + (float)blue * 0.11f);
+		uint32_t gray = (uint32_t) ((float)red * 0.30f + (float)green * 0.59f + (float)blue * 0.11f);
 
 		// 7. Now we need to put the pixel data back into the image's data.
 		// We do the opposite of how we extracted the ARGB values above and
