+++ tuxcap/include/Quantize.h
@@ -6,7 +6,7 @@
 namespace Sexy
 {
 
-bool Quantize8Bit(const ulong* theSrcBits, int theWidth, int theHeight, uchar* theDestColorIndices, ulong* theDestColorTable);
+bool Quantize8Bit(const uint32_t* theSrcBits, int theWidth, int theHeight, uchar* theDestColorIndices, uint32_t* theDestColorTable);
 
 }
