+++ tuxcap/include/SexyAppBase.h
@@ -341,7 +341,7 @@ public:
 	Uint32					mTimeLoaded;
 	SexyString				mTitle;	
 	Image*					mCursorImages[NUM_CURSORS];
-        ulong mFPSStartTick;
+        uint32_t mFPSStartTick;
 	Buffer					mDemoBuffer;
 	MemoryImageSet			mMemoryImageSet;
 	SharedImageMap			mSharedImageMap;
@@ -418,10 +418,10 @@ public:
 	void					MirrorImage(Image* theImage);
 	void					FlipImage(Image* theImage);
 	void					RotateImageHue(Sexy::MemoryImage *theImage, int theDelta);
-	ulong					HSLToRGB(int h, int s, int l);
-	ulong					RGBToHSL(int r, int g, int b);
-	void					HSLToRGB(const ulong* theSource, ulong* theDest, int theSize);
-	void					RGBToHSL(const ulong* theSource, ulong* theDest, int theSize);
+	uint32_t					HSLToRGB(int h, int s, int l);
+	uint32_t					RGBToHSL(int r, int g, int b);
+	void					HSLToRGB(const uint32_t* theSource, uint32_t* theDest, int theSize);
+	void					RGBToHSL(const uint32_t* theSource, uint32_t* theDest, int theSize);
 	void					PrecacheAdditive(MemoryImage* theImage);
 	void					PrecacheAlpha(MemoryImage* theImage);
 	void					PrecacheNative(MemoryImage* theImage);
@@ -466,13 +466,13 @@ public:
 	bool					RegistryReadString(const std::string& theValueName, std::string* theString);
 	bool					RegistryReadInteger(const std::string& theValueName, int* theValue);
 	bool					RegistryReadBoolean(const std::string& theValueName, bool* theValue);
-	bool					RegistryReadData(const std::string& theValueName, uchar* theValue, ulong* theLength);
+	bool					RegistryReadData(const std::string& theValueName, uchar* theValue, uint32_t* theLength);
 	bool					RegistryEraseKey(const SexyString& theKeyName);
 	void					RegistryEraseValue(const SexyString& theValueName);
 	bool					RegistryWriteString(const std::string& theValueName, const std::string& theString);
 	bool					RegistryWriteInteger(const std::string& theValueName, int theValue);
 	bool					RegistryWriteBoolean(const std::string& theValueName, bool theValue);
-	bool					RegistryWriteData(const std::string& theValueName, const uchar* theValue, ulong theLength);	
+	bool					RegistryWriteData(const std::string& theValueName, const uchar* theValue, uint32_t theLength);	
 	virtual void			WriteToRegistry();
 	virtual void			ReadFromRegistry();
 	SexyString				GetString(const std::string& theId);
@@ -506,9 +506,9 @@ public:
 
 protected:	
 	// Registry helpers
-	bool					RegistryRead(const std::string& theValueName, ulong* theType, uchar* theValue, ulong* theLength);
-	bool					RegistryReadKey(const std::string& theValueName, ulong* theType, uchar* theValue, ulong* theLength, HKEY theMainKey = HKEY_CURRENT_USER);
-	bool					RegistryWrite(const std::string& theValueName, ulong theType, const uchar* theValue, ulong theLength);
+	bool					RegistryRead(const std::string& theValueName, uint32_t* theType, uchar* theValue, uint32_t* theLength);
+	bool					RegistryReadKey(const std::string& theValueName, uint32_t* theType, uchar* theValue, uint32_t* theLength, HKEY theMainKey = HKEY_CURRENT_USER);
+	bool					RegistryWrite(const std::string& theValueName, uint32_t theType, const uchar* theValue, uint32_t theLength);
 
 	virtual bool			DoUpdateFrames();
 	virtual void			DoUpdateFramesF(float theFrac);
@@ -524,7 +524,7 @@ protected:
 	virtual void			DeleteNativeImageData();	
 	virtual void			PostDDInterfaceInitHook();
 	virtual void			PreDisplayHook();
-	bool					WriteBytesToFile(const std::string& theFileName, const void *theData, unsigned long theDataLen);
+	bool					WriteBytesToFile(const std::string& theFileName, const void *theData, uint32_t theDataLen);
 	void					RestoreScreenResolution();
 	void					DoExit(int theCode);
 	virtual bool			ChangeDirHook(const char *theIntendedPath);
