+++ tuxcap/lib/MTRand.cpp
@@ -66,7 +66,7 @@ MTRand::MTRand(const std::string& theSerialData)
 	mti=MTRAND_N+1; /* mti==MTRAND_N+1 means mt[MTRAND_N] is not initialized */
 }
 
-MTRand::MTRand(unsigned long seed)    
+MTRand::MTRand(uint32_t seed)    
 {
 	SRand(seed);
 }
@@ -100,7 +100,7 @@ void MTRand::SRand(const std::string& theSerialData)
 		SRand(4357);
 }
 
-void MTRand::SRand(unsigned long seed)
+void MTRand::SRand(uint32_t seed)
 {
 	if (seed == 0)
 		seed = 4357;
@@ -123,16 +123,16 @@ void MTRand::SRand(unsigned long seed)
 	}
 }
 
-unsigned long MTRand::Next()
+uint32_t MTRand::Next()
 {
 	assert(gRandAllowed==0);
 	return NextNoAssert();
 }
 
-unsigned long MTRand::NextNoAssert()
+uint32_t MTRand::NextNoAssert()
 {
-    unsigned long y;
-    static unsigned long mag01[2]={0x0, MATRIX_A};
+    uint32_t y;
+    static uint32_t mag01[2]={0x0, MATRIX_A};
     /* mag01[x] = x * MATRIX_A  for x=0,1 */
 
 	if (mti >= MTRAND_N) { /* generate MTRAND_N words at one time */
@@ -167,12 +167,12 @@ unsigned long MTRand::NextNoAssert()
     return y; 
 }
 
-unsigned long MTRand::NextNoAssert(unsigned long range)
+uint32_t MTRand::NextNoAssert(uint32_t range)
 {
 	return NextNoAssert() % range;
 }
 
-unsigned long MTRand::Next(unsigned long range)
+uint32_t MTRand::Next(uint32_t range)
 {
 	assert(gRandAllowed==0);
 	return NextNoAssert( range );
